Refactor sort module: fuse passes, share comparators, host builder
Addresses three review comments from @inulty-dfe on
version-1/refactor:

1. src/sort/index.js:22-25 - Collapse the `.map(normaliseFn).map(weight)`
   pair into a single pass. `normaliseFn` already mutates and returns
   the option, so the second iteration was redundant.

2. src/sort/comparators.js (new) - Extract `hasWeight`,
   `byWeightThenAlphabetically`, and `optionName` so both
   `sort/index.js` and `sort/search-index.js` can import them without
   introducing a circular dependency once the builder moves in.

3. src/dfe-autocomplete.js - Move `buildSortFunction` into
   `src/sort/index.js`. The helper exists purely to assemble a sort
   function from `createSort`/`createIndexedSort`/`createRemoveStopWords`,
   so it belongs alongside the rest of the sort module. The autocomplete
   entry point now only imports `buildSortFunction`; the
   `createRemoveStopWords` and `createIndexedSort` imports it used to
   pull in for this function only are gone.

All 262 existing tests still pass.
