+++ src/sort/index.js
@@ -1,26 +1,17 @@
 import defaultClean from './clean'
 import defaultCalculateWeight from './calculateWeight'
 import { normalise, createNormaliser } from './cleanse'
-
-const hasWeight = (option) => option.weight > 0
-
-const byWeightThenAlphabetically = (a, b) => {
-  if (a.weight > b.weight) return -1
-  if (a.weight < b.weight) return 1
-  if (a.name < b.name) return -1
-  if (a.name > b.name) return 1
-  return 0
-}
-
-const optionName = (option) => option.name
+import { hasWeight, byWeightThenAlphabetically, optionName } from './comparators'
+import { createRemoveStopWords } from './stop_words'
+import { createIndexedSort } from './search-index'
 
 function buildPipeline (cleanFn, normaliseFn, calculateWeightFn) {
   return (query, options) => {
     const cleanQuery = cleanFn(query)
 
     return options
-      .map(normaliseFn)
       .map((option) => {
+        normaliseFn(option)
         option.weight = calculateWeightFn(option.clean, cleanQuery) * option.clean.boost
         return option
       })
@@ -39,5 +30,26 @@ export function createSort ({ clean, removeStopWords, calculateWeight } = {}) {
   return buildPipeline(cleanFn, normaliseFn, calculateWeightFn)
 }
 
+const defaultSort = buildPipeline(defaultClean, normalise, defaultCalculateWeight)
+
+export function buildSortFunction (options, libraryOptions) {
+  if (libraryOptions.sort) return libraryOptions.sort
+
+  const hasCustomInternals = libraryOptions.stopWords || libraryOptions.calculateWeight || libraryOptions.clean
+  if (!hasCustomInternals && !libraryOptions.useSearchIndex) return defaultSort
+
+  const customConfig = {
+    clean: libraryOptions.clean,
+    removeStopWords: libraryOptions.stopWords ? createRemoveStopWords(libraryOptions.stopWords) : undefined,
+    calculateWeight: libraryOptions.calculateWeight
+  }
+
+  if (libraryOptions.useSearchIndex) {
+    return createIndexedSort(options, customConfig)
+  }
+
+  return createSort(customConfig)
+}
+
 export { normalise, hasWeight, byWeightThenAlphabetically, optionName }
-export default buildPipeline(defaultClean, normalise, defaultCalculateWeight)
+export default defaultSort
